# Remove VirtualTable matrix rotation