Instead of rotating the generated data from row to columns, access each row/column directly. Although there were optimizations around space within the rotation, this improves query performance by a factor of `1.25`.

Before:
```
Benchmark                         Time(ns)    CPU(ns) Iterations
----------------------------------------------------------------
SQL_virtual_table_registry            2134       2128     311644                                 
SQL_virtual_table_internal          113019     112584       6089                                 
SQL_virtual_table_internal_long    1971752    1964762        340                                 
SQL_virtual_table_internal_wide    2192625    2185087        310                                 
SQL_select_metadata                  40653      40457      16772                                 
SQL_select_basic                    113714     113172       5979  
```

After:
```
Benchmark                         Time(ns)    CPU(ns) Iterations
----------------------------------------------------------------
SQL_virtual_table_registry            2159       2148     315375                                 
SQL_virtual_table_internal          112635     112197       6186                                 
SQL_virtual_table_internal_long    1595019    1587392        424                                 
SQL_virtual_table_internal_wide    1825092    1818562        368                                 
SQL_select_metadata                  41246      41101      16989                                 
SQL_select_basic                    106747     106351       6162 
```