@theopolis 2015-08-18
I think this will flatten the daemon's memory footprint a little. During the EOF table generation the column-sorted (rotated matrix) data was not cleared.